メニューの部分をコンポーネント化して外だししました。
